cmd/builder: automatically choose nacltest.bash for nacl- builds
This is a recapitulation of CL 114780043.

This change extends the buildCmd logic to automatically choose the
nacltest.bash build script if the builder key represents a nacl job.

This is a change I have been carrying for over year now which I have
to remember to reapply every time I rebuild the nacl-arm builder.

Change-Id: I3e86a3e62671cda6feb318bb4b312f5462436f27
Reviewed-on: https://go-review.googlesource.com/4261
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>
Reviewed-by: Andrew Gerrand <adg@golang.org>