cmd/builder: do not set GOROOT_FINAL
Fixes #9820

The builders executed a build in an environment which used the
GOROOT_FINAL override to set the final destination of the build
to be `/usr/local/go`.

This behaviour was carried over from the time when the builders
themselves produced the binary releases. It is not relevant now.

https://codereview.appspot.com/2126042/diff/12008/misc/dashboard/builder/main.go#newcode238

Because GOROOT_FINAL was set, the value of GOROOT embedded in the
runtime was incorrect, causing the path to the gdb python helper
to also be incorrect and the test to fail (unless you had a version
of Go in /usr/local/go).

Change-Id: I5d4b34931ba171ff97a22b95cbeabdf6a0eb304d
Reviewed-on: https://go-review.googlesource.com/4300
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>
Reviewed-by: Minux Ma <minux@golang.org>