env: update Plan 9 image
This new image provides a new tool, called
aux/randfs. This is a pseudo-random file
system which can be mounted on top of /dev/random.

We noticed the Go binaries on GCE were running
much slower than usual. It happened because the
runtime was regularly waiting after /dev/random,
which wasn't able to produce random bytes as
fast as required.

A simple workaround was to write a pseudo-random
file system, acting as /dev/random and initialized
from a true random source. It will be able
to produce much more bytes per second than
the random generator provided by the kernel.

Update golang/go#9491

Change-Id: I81158adef7332d10295c2245b4564ae18ebb3a14
Reviewed-on: https://go-review.googlesource.com/6170
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>