cl: hide DoNotReview changes
The actual spelling of DoNotReview that has meaning
to the dashboard includes spaces, but I can't put that
string in the commit message because it makes the
CL uncommitable. :-)

Add dnr flag so people can ask for them.

If people do ask, show DoNotReview as the reviewer.
Similarly, show Closed as the reviewer for closed CLs.

Change-Id: Ibe758bbb5f5f03fcb5035e7cb068c6d8d0b6baa4
Reviewed-on: https://go-review.googlesource.com/11260
Reviewed-by: Andrew Gerrand <adg@golang.org>
Run-TryBot: Russ Cox <rsc@golang.org>