cmd/coordinator: don't reuse main buildlet when broken, make timeouts fatal
This CL makes the main buildlet and any helper buildlets all
equivalent once we get to the final stage of the build (test
execution). Now the death of the main buildlet isn't critical and the
testing continues as long as any buildlet is alive (or any helper
buildlet is on its way).

It also makes timeouts (20 minutes) of remote commands fatal, now that
health checking is reliable for all buildlet types.

Fixes golang/go#12685

Change-Id: I07d1e1fabf1027ee24ce679e4fcc8bd09bed6d83
Reviewed-on: https://go-review.googlesource.com/14774
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>