cmd/fetchlogs: put logs in a standard directory
Currently fetchlogs downloads logs to a directory structure created in
the current directory. This makes it difficult to run (you have to
make sure you're in the right directory) and makes it hard for other
tools (e.g., log processing tools) to find these logs.

This change adds a -dir flag to fetchlogs to specify a download
directory. The default value of this flag is derived from the cache
directory of the XDG Base Directory Specification.

Change-Id: I05f01a3db7423250cfef18c42ff38b599b097b35
Reviewed-on: https://go-review.googlesource.com/16818
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>