dashboard: use Go 1.6 to bootstrap Go on Plan 9
On the Plan 9 builder, we are experiencing regular
failures in cmd/compile when building go_boostrap.

We suspect the failures to be related to a bug in
the Go 1.4 runtime. This is affecting the cmd/compile
program, since it has been built with Go 1.4 and
linked with the Go 1.4 runtime.

Updates golang/go#14650

Change-Id: I78729c8660314be381d7b668810c6f64214f2358
Reviewed-on: https://go-review.googlesource.com/21411
Reviewed-by: Richard Miller <millerresearch@gmail.com>
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>