all: import godash into the build repo
This imports the godash command-line tool from rsc.io/github/godash. The
tool is refactored into a reusable library, in preparation for running
the tool as part of the build.golang.org App Engine app.

The behavior is mostly the same as the version on rsc.io, but it
eliminates the ability to use a data cache on disk, and it replaces the
hardcoded release version with logic that uses the lowest version
milestone that is marked open on Github.

Change-Id: If5ca5483a07ec60f74f45d140974c4140fff34ca
Reviewed-on: https://go-review.googlesource.com/23095
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>