cmd/godash: load reviewer information from Github
This allows the godash command to function without a local git clone of
the go repo. Because that is slow, it also allows the godash command to
cache commit information so it only needs to fetch new commits on each
run.

This behavior change is necessary so that we can run godash on App
Engine, which does not have a local filesystem.

Change-Id: I5bcdae62b22d46c0d6e02a152f58fd3c110e7a82
Reviewed-on: https://go-review.googlesource.com/23139
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>