devapp, godash: collect issue statistics for plotting
A subsequent CL will add a dashboard for graphing project health; this
CL collects information about when issues are opened, closed, and
milestoned.

This also includes a refactor of the guts of devapp to make it more
modular.

It also adds logging to the godash library, which can be activated with
the -v flag in the godash command.

Change-Id: I54c1419435e496f5c0e2e7f4b966b3ee4de0b0a1
Reviewed-on: https://go-review.googlesource.com/28091
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>