cmd/coordinator, dashboard: remove some trybots, shard others wider
I'm aiming to have trybot runs finish in under 5 minutes.

This CL removes openbsd-386-gce58 and freebsd-386-gce101 from the trybot set.

openbsd-386-gce58 is the slowest builder. It has an average speed of
722 seconds (and 95 percentile of 923 seconds) over the past week, and
that's sharded over 4 machines. Too slow. It's not worth the resources
to keep it as a trybot. It hasn't caught any interesting bugs. This
builder will still run, but not as a pre-submit trybot.

freebsd-386-gce101 is not slow, but we're removing it to shift its
resources to shard other builders wider.

The coordinator now supports varying the build sharding width based on
whether a build is for a trybot or not. This CL defines separate
numbers for each, sharding builds wider as needed for some trybots.

freebsd-amd64-gce101 goes from 4 to 5 machines in try runs, and down
to 3 when not in try runs.

linux-amd64-race gets one more machine during try runs, and one fewer
in regular runs.

linux-arm goes from 7 machines always, to 3 or 8, depending on whether
it's a try run.

openbsd-amd64-58 goes from 4 to 3 or 6.

windows-amd64-gce goes from 4 to 2 or 6.

windows-amd64-race goes from 4 to 2 or 6.

darwin-amd64-10_11 goes from 3 to 3 or 4.

I'll see how these do over the next few days and readjust as needed.

Also in this CL: fix the constants for the expected duration of
make.bash, which impact when we schedule the creation of test sharding
helper buildlets. We were creating them too early before, wasting
resources.

Change-Id: I38a9b24841e196f1eb668de058c49af8c1d1c64f
Reviewed-on: https://go-review.googlesource.com/29116
Reviewed-by: Quentin Smith <quentin@golang.org>