cmd/release: notify explorer of PATH update after install on Win
Windows installer will not broadcast WM_SETTINGCHANGE when a reboot is
pending. This message is required for explorer.exe to reload the PATH
from the registry so future cmd.exe processes will launch with go/bin in
the PATH.

Use the Wix extension to always broadcast this message after install.

More info:
http://wixtoolset.org/documentation/manual/v3/customactions/wixsettingchange.html

Fixes golang/go#18680

Change-Id: I4f3af49f167f3875bbf8e898f4d50ee358b4e35d
Reviewed-on: https://go-review.googlesource.com/36324
Reviewed-by: Chris Broadfoot <cbro@golang.org>