cmd/coordinator: add tests for buildlet create
Specifically test that the context is not canceled when the create
function returns, which is the case that bit us yesterday. This
still fails "go vet" and we should figure out how to pass the cancel
function down or otherwise restructure the code.

Add the beginnings of a buildlet test framework, which will hopefully
make it easier to write more tests without needing to start up a
cluster or integrate with GCE. It would be nice if buildlet.Client
was an interface, not a concrete type, so we could store various
pieces of information on it and then test for them.

Change-Id: I75acc3252b59110647a8ad8d8f7c7cce66598a17
Reviewed-on: https://go-review.googlesource.com/37146
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>