devapp: fix travis-ci tests
These tests pass locally (on my Mac) because Go builds in
/etc/apache2/mime.types, which maps the .ico file extension to the
image/x-icon content-type, which we check for in the test.

The Travis-CI builders don't have that file, so they return
vnd.microsoft.icon, which makes the test fail. Instead let's hard code
the content-type that we want, so the test passes.

Verified this failed by moving /etc/apache2/mime.types on my Mac,
observing a failure, then fixing the test and observing a pass.

Change-Id: I42d07fe413d2686d186f9157af8988e3a146a877
Reviewed-on: https://go-review.googlesource.com/38440
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>