cmd/coordinator: allow fetching logs of in-progress try sets
statusDone contains the last maxStatusDone (currently 30) logs, but
logs for in-progress try sets are still available via the trySet until
the trySet is complete.

Currently, if you go to the main status page of the farmer, there is
an "Active Trybot Runs" section at the top. If you click on the status
link (taking you to /try?commit=xxx), it shows partial logs for each
of the buildlets. However, if you click the /temporarylogs link for
one of those buildlets, you get a 404 if that buildStatus has fallen
off the end of maxStatusDone. This CL makes it so that any
/temporarylogs link you can get from a /try page will always contain
the actual log, which we were already storing in memory anyway.

Change-Id: Ia07b9c8b1a16da23936b9b32e39827e591e3e471
Reviewed-on: https://go-review.googlesource.com/39312
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>