cmd/coordinator: fix local development environment
Previously the server would fail to start if a valid token could not
be found. The local development environment still doesn't do much, but
it should let you view and edit the HTML and confirm that the server
starts, so ignore the error if we are running in dev mode.

Add a short README explaining how to start and view the coordinator
server locally.

Fixes golang/go#19828.
Fixes golang/go#18291.

Change-Id: I91a3ce49e1e9ea18ca19f3867edb2c71fc1b5124
Reviewed-on: https://go-review.googlesource.com/39297
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>