maintner: don't use github caching transport after getting webhook update
If we got a webhook update, we know there's new data. We're not in
polling mode, so don't use the caching transport.

The caching transport was being too agressive and we were missing
label changes to issues, since the Issues.ListByRepo call to get
issues sorted by mod time descending was being cached.

This speeds up gopherbot. Gopherbot can react in realtime to changes,
but only if they make it into maintner in realtime.

Updates golang/go#20819

Change-Id: I5c0a0bbcb4bf0d9d484b9279ff5a80732b4e1269
Reviewed-on: https://go-review.googlesource.com/47133
Reviewed-by: Andrew Bonventre <andybons@google.com>