all: normalize GKE configs/Makefiles to follow common patterns
When a push is made, the latest image is pushed to the registry
with the git hash of the repo and also the same image is pushed
with the :latest tag. This ensures that :latest always points to
the most recent image and that kubectl set image will trigger an
update on the pods without having to delete them.

Fixes golang/go#21772

Change-Id: I97c893f181cdd8891eb87efb3f9ab42ede107d97
Reviewed-on: https://go-review.googlesource.com/61730
Reviewed-by: Sarah Adams <shadams@google.com>