cmd/gitmirror: send new branch commits to dashboard
When a new branch is created, we will have already seen the base
commit (last commit on branch from which new branch is forked).
When this happens, we have been inserting a copy of the commit
as the base commit, and not the pointer we have already in our
map of commits on a branch.

This messes up the linked list (we now have two copies of a commit
in the list).

This fix ensures we only ever keep one copy of a commit in our set
of known commits.

Fixes golang/go#21555

Change-Id: I8cb5023cf9dd8fee82a86346b940bdd67b0ce8cc
Reviewed-on: https://go-review.googlesource.com/64150
Reviewed-by: Andrew Bonventre <andybons@golang.org>