cmd/release: allow uploading artifacts from GCS
Use gs:// URIs to pass in individual files, or "/..." in the path to
use all files in the tree. SHA256 checksums are pre-computed, so the
files can simply be copied between GCS buckets rather than being
downloaded to the local filesystem first.

Input files are deduplicated by basename, and signed artifacts are preferred.
They are signified by "/signed/" being in the GCS path.

This means that a command like:

$ release -upload gs://golang-release-staging/go1.9.2rc2/...

will upload all of the artifacts, including those in the "signed"
subdirectory.

Change-Id: I088e4b781d1f3429cdff2b524a1706ab55013b67
Reviewed-on: https://go-review.googlesource.com/73070
Reviewed-by: Russ Cox <rsc@golang.org>