cmd/coordinator, internal/buildstats: fix buildstats sync, update deps
Around Oct 20th, we stopped syncing build stats to BigQuery.

Either the client library was updated around Oct 20th, or the server
changed behavior.

In any case, update the deps (to get a newer version of the bigquery
package) and update the code to work with the new SQL syntax.

This also fixes a bogus loop that was harmless but made the sync
slower than it could've been. (The logic was correct in the nearly
identical span sync later in the file.)

Also adds some TODOs for other log spam I see.

Change-Id: I2aa73084aeb19378bf260839555103fb48eac168
Reviewed-on: https://go-review.googlesource.com/78455
Reviewed-by: Andrew Bonventre <andybons@golang.org>