cmd/gerritbot: ensure commit message consistency
The message generated by the commitMessage method
does not always match the one that Gerrit sees.
This is because the commit-msg hook runs git stripspace
and has different logic for placing the Change-Id.

Remove the commit-msg hook and insert a Change-Id
(a SHA-1 hash preceded with an I) using the Pull
Request’s ID and run git stripspace on the generated
commit message.

This is preparation for updating the Gerrit CL once
the title or description changes.

Also updates Dockerfile.0 to use 1.10 and disable
cgo so that building doesn’t emit errors.

Updates golang/go#23756

Change-Id: Ib437d3e9c1496ad1bee6268afcbe7c766eb82bde
Reviewed-on: https://go-review.googlesource.com/96995
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>