cmd/gerritbot: use full commit message to determine staleness
Instead of checking only against the last revision of a GitHub PR,
check to see if the Gerrit commit message matches that of the one
determined by the PR, which is the source of truth due to
GerritBot’s one-way nature.

Also deletes some dead code and adds a dry-run flag.

Fixes golang/go#23756

Change-Id: Ib2f93ebaf3f6af274f62d0e6b1e8bcab89fb7109
Reviewed-on: https://go-review.googlesource.com/96999
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>