cmd/releasebot: use comment ID in updateGithubComment
When editing a comment, the current code used the issue number in place
of the comment ID. But the EditComment endpoint needs the comment ID.

The issue number is not needed at all in updateGithubComment because the
EditComment endpoint uses the repository and comment ID to uniquely
identify the comment.

References:

-	https://developer.github.com/v3/issues/comments/#edit-a-comment.
-	https://godoc.org/github.com/google/go-github/github#IssuesService.EditComment.

Updates golang/go#24598.
Updates google/go-github#883.
Updates google/go-github#597.

Change-Id: Iae9d967d7be7a75b1bcee7118a3c80fe8f2375b4
Reviewed-on: https://go-review.googlesource.com/103398
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>